Fix log initialization and formatting
This commit fixes two small issues in the abseil logger. First, a call to the
log initialization function was missing, which would lead to verbose warnings
about uninitialized logging. Second, the formatting of the log messages would
include a long prefix, which was not ideal, especially when running the
benchmarks script.